refactor(quarkifier): clean architecture, remove dead code, introduce subpackages
