+++ quarkifier/src/main/java/com/clementguillot/quarkifier/dev/DevModeLauncher.java
@@ -1,5 +1,9 @@
-package com.clementguillot.quarkifier;
+package com.clementguillot.quarkifier.dev;
 
+import com.clementguillot.quarkifier.AugmentationException;
+import com.clementguillot.quarkifier.BuildProperties;
+import com.clementguillot.quarkifier.QuarkifierConfig;
+import com.clementguillot.quarkifier.maven.MavenCoordinateParser;
 import io.quarkus.bootstrap.BootstrapConstants;
 import io.quarkus.bootstrap.app.QuarkusBootstrap;
 import io.quarkus.bootstrap.model.ApplicationModel;
@@ -16,8 +20,11 @@ import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.jar.Attributes;
+import java.util.jar.JarFile;
 import java.util.jar.Manifest;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
@@ -27,9 +34,9 @@ import java.util.zip.ZipOutputStream;
  *
  * <p>Following Maven's {@code DevMojo} approach, this launcher creates a minimal "dev jar"
  * containing the serialized {@link DevModeContext} and a manifest classpath pointing to the
- * deployment jars. A <b>separate JVM process</b> is started with {@code java -jar dev.jar},
- * which invokes {@link DevModeMain#main} to bootstrap {@code IsolatedDevModeMain} inside a
- * clean augment classloader with no overlap from the quarkifier's system classloader.
+ * deployment jars. A <b>separate JVM process</b> is started with {@code java -jar dev.jar}, which
+ * invokes {@link DevModeMain#main} to bootstrap {@code IsolatedDevModeMain} inside a clean augment
+ * classloader with no overlap from the quarkifier's system classloader.
  */
 public final class DevModeLauncher {
 
@@ -45,22 +52,26 @@ public final class DevModeLauncher {
   public static void launch(QuarkifierConfig config, ApplicationModel appModel)
       throws AugmentationException {
     try {
-      DevModeContext context = buildDevModeContext(config, appModel);
+      DevModeContext context = buildDevModeContext(config);
 
       // 1. Serialize the ApplicationModel to a temp file
       Path serializedModel = BootstrapUtils.serializeAppModel(appModel, false);
 
       // 2. Create the dev jar (like Maven's DevMojo / DevModeCommandLineBuilder)
-      Path devJar = createDevJar(context, config.deploymentClasspath(),
-          config.applicationClasspath());
+      Path devJar =
+          createDevJar(
+              context, config.deploymentClasspath(), config.applicationClasspath());
 
       // 3. Build the java command
       String javaBin = System.getProperty("java.home") + "/bin/java";
       List<String> cmd = new ArrayList<>();
       cmd.add(javaBin);
       cmd.add("-Djava.util.logging.manager=org.jboss.logmanager.LogManager");
-      cmd.add("-D" + BootstrapConstants.SERIALIZED_APP_MODEL + "="
-          + serializedModel.toAbsolutePath());
+      cmd.add(
+          "-D"
+              + BootstrapConstants.SERIALIZED_APP_MODEL
+              + "="
+              + serializedModel.toAbsolutePath());
       cmd.add("-jar");
       cmd.add(devJar.toAbsolutePath().toString());
 
@@ -84,48 +95,29 @@ public final class DevModeLauncher {
   }
 
   /**
-   * Creates a minimal JAR containing the serialized {@link DevModeContext} and a manifest
-   * with {@code Class-Path} entries pointing to the deployment jars.
+   * Creates a minimal JAR containing the serialized {@link DevModeContext} and a manifest with
+   * {@code Class-Path} entries pointing to the deployment jars.
    *
-   * <p>The manifest classpath includes all deployment jars EXCEPT runtime Quarkus extension
-   * jars (those with {@code META-INF/quarkus-extension.properties}). Runtime extension jars
-   * like {@code quarkus-arc}, {@code quarkus-rest}, {@code smallrye-config-inject} contain
-   * CDI beans whose ArC-generated proxies cause VerifyError when the bean class is loaded
-   * by the system CL and the proxy by the augment/runtime CL.
+   * <p>The manifest classpath includes all deployment jars EXCEPT runtime Quarkus extension jars
+   * (those with {@code META-INF/quarkus-extension.properties}). Runtime extension jars like {@code
+   * quarkus-arc}, {@code quarkus-rest}, {@code smallrye-config-inject} contain CDI beans whose
+   * ArC-generated proxies cause VerifyError when the bean class is loaded by the system CL and the
+   * proxy by the augment/runtime CL.
    */
-  private static Path createDevJar(DevModeContext context,
-      List<Path> deploymentClasspath, List<Path> applicationClasspath) throws Exception {
+  private static Path createDevJar(
+      DevModeContext context, List<Path> deploymentClasspath, List<Path> applicationClasspath)
+      throws Exception {
     Path tempFile = Files.createTempFile("quarkus-dev", ".jar");
     tempFile.toFile().deleteOnExit();
 
-    // Identify runtime extension jars (those with quarkus-extension.properties).
-    // These must NOT be on the manifest classpath — they are loaded exclusively
-    // by the augment/runtime classloader from the ApplicationModel.
-    java.util.Set<String> extensionArtifactIds = new java.util.HashSet<>();
-    for (Path jar : applicationClasspath) {
-      try (var jf = new java.util.jar.JarFile(jar.toFile())) {
-        if (jf.getEntry("META-INF/quarkus-extension.properties") != null) {
-          extensionArtifactIds.add(MavenCoordinateParser.parse(jar).artifactId());
-        }
-      } catch (Exception ignored) {
-        // Skip jars that can't be read
-      }
-    }
-    // Also exclude jars whose classes are loaded by both the system CL and the
-    // QuarkusClassLoader, causing ClassCastException across classloader boundaries.
-    extensionArtifactIds.add("smallrye-config-inject");
-    extensionArtifactIds.add("smallrye-config");
-    extensionArtifactIds.add("smallrye-config-core");
-    extensionArtifactIds.add("smallrye-config-common");
-    extensionArtifactIds.add("microprofile-config-api");
+    Set<String> extensionArtifactIds = collectExtensionArtifactIds(applicationClasspath);
 
     try (ZipOutputStream out = new ZipOutputStream(new FileOutputStream(tempFile.toFile()))) {
       out.putNextEntry(new ZipEntry("META-INF/"));
 
       Manifest manifest = new Manifest();
       manifest.getMainAttributes().put(Attributes.Name.MANIFEST_VERSION, "1.0");
-      manifest.getMainAttributes().put(Attributes.Name.MAIN_CLASS,
-          DevModeMain.class.getName());
+      manifest.getMainAttributes().put(Attributes.Name.MAIN_CLASS, DevModeMain.class.getName());
 
       // Include all deployment jars except runtime extension jars
       StringBuilder classPath = new StringBuilder();
@@ -136,7 +128,9 @@ public final class DevModeLauncher {
           classPath.append(uri).append(" ");
         }
       }
-      manifest.getMainAttributes().put(Attributes.Name.CLASS_PATH, classPath.toString().trim());
+      manifest
+          .getMainAttributes()
+          .put(Attributes.Name.CLASS_PATH, classPath.toString().trim());
 
       out.putNextEntry(new ZipEntry("META-INF/MANIFEST.MF"));
       manifest.write(out);
@@ -153,9 +147,35 @@ public final class DevModeLauncher {
   }
 
   /**
-   * Builds a {@link DevModeContext} configured for Bazel-managed dev mode.
+   * Collects artifact IDs that must be excluded from the dev jar's manifest classpath.
+   *
+   * <p>This includes all runtime extension jars (detected via {@code
+   * quarkus-extension.properties}) plus a hardcoded set of SmallRye Config jars. These are loaded
+   * exclusively by the augment/runtime classloader; putting them on the system classpath causes
+   * ClassCastException and VerifyError across classloader boundaries.
    */
-  static DevModeContext buildDevModeContext(QuarkifierConfig config, ApplicationModel appModel) {
+  private static Set<String> collectExtensionArtifactIds(List<Path> applicationClasspath) {
+    Set<String> extensionArtifactIds = new HashSet<>();
+    for (Path jar : applicationClasspath) {
+      try (var jf = new JarFile(jar.toFile())) {
+        if (jf.getEntry("META-INF/quarkus-extension.properties") != null) {
+          extensionArtifactIds.add(MavenCoordinateParser.parse(jar).artifactId());
+        }
+      } catch (Exception ignored) {
+      }
+    }
+    // Also exclude SmallRye Config jars — they are loaded by both the system CL and the
+    // QuarkusClassLoader, causing ClassCastException across classloader boundaries.
+    extensionArtifactIds.add("smallrye-config-inject");
+    extensionArtifactIds.add("smallrye-config");
+    extensionArtifactIds.add("smallrye-config-core");
+    extensionArtifactIds.add("smallrye-config-common");
+    extensionArtifactIds.add("microprofile-config-api");
+    return extensionArtifactIds;
+  }
+
+  /** Builds a {@link DevModeContext} configured for Bazel-managed dev mode. */
+  static DevModeContext buildDevModeContext(QuarkifierConfig config) {
     var context = new DevModeContext();
     context.setAbortOnFailedStart(true);
     context.setLocalProjectDiscovery(false);
@@ -165,10 +185,8 @@ public final class DevModeLauncher {
     context.setProjectDir(config.outputDir().toAbsolutePath().toFile());
 
     // Platform properties for SmallRye Config expression resolution
-    context.getBuildSystemProperties().put(
-        "platform.quarkus.native.builder-image",
-        "quay.io/quarkus/ubi-quarkus-mandrel-builder-image:jdk-21");
-    context.getBuildSystemProperties().put("quarkus.package.jar.type", "fast-jar");
+    BuildProperties.defaults().forEach(
+        (k, v) -> context.getBuildSystemProperties().put((String) k, (String) v));
 
     // Build ModuleInfo for the application root
     Path appJar = config.applicationClasspath().get(0);
