+++ quarkifier/src/main/java/com/clementguillot/quarkifier/extension/ExtensionScanner.java
@@ -1,4 +1,4 @@
-package com.clementguillot.quarkifier;
+package com.clementguillot.quarkifier.extension;
 
 import java.io.IOException;
 import java.io.InputStream;
