+++ quarkifier/src/test/java/com/clementguillot/quarkifier/extension/VersionCheckerTest.java
@@ -1,4 +1,4 @@
-package com.clementguillot.quarkifier;
+package com.clementguillot.quarkifier.extension;
 
 import static org.junit.jupiter.api.Assertions.*;
